efi/cper, cxl: Make definitions and structures global
commit 958c3a6706863f9f77b21c90d2428473441cd8a1 upstream.

In preparation to add tracepoint support, move protocol error UUID
definition to a common location, Also, make struct CXL RAS capability,
cxl_cper_sec_prot_err and CPER validation flags global for use across
different modules.

Link: https://patch.msgid.link/20250123084421.127697-3-Smita.KoralahalliChannabasappa@amd.com
